backend/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
Django==4.2.4
django-graphql-jwt==0.4.0
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
promise==2.3
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
